+++ lib/src/marker_url.dart
@@ -1,7 +1,9 @@
 import 'dart:io';
+
 import 'package:map_launcher/src/models.dart';
 import 'package:map_launcher/src/utils.dart';
 
+/// Returns a url that is used by [showMarker]
 String getMapMarkerUrl({
   required MapType mapType,
   required Coords coords,
@@ -16,7 +18,8 @@ String getMapMarkerUrl({
       return Utils.buildUrl(
         url: Platform.isIOS ? 'comgooglemaps://' : 'geo:0,0',
         queryParams: {
-          'q': '${coords.latitude},${coords.longitude}($title)',
+          'q':
+              '${coords.latitude},${coords.longitude}${title != null && title.isNotEmpty ? '($title)' : ''}',
           'zoom': '$zoomLevel',
           ...(extraParams ?? {}),
         },
@@ -26,7 +29,8 @@ String getMapMarkerUrl({
       return Utils.buildUrl(
         url: 'http://maps.google.com/maps',
         queryParams: {
-          'q': '${coords.latitude},${coords.longitude}($title)',
+          'q':
+              '${coords.latitude},${coords.longitude}${title != null && title.isNotEmpty ? '($title)' : ''}',
           'zoom': '$zoomLevel',
           ...(extraParams ?? {}),
         },
@@ -186,6 +190,139 @@ String getMapMarkerUrl({
             'https://share.here.com/l/${coords.latitude},${coords.longitude},$title',
         queryParams: {
           'z': '$zoomLevel',
+          ...(extraParams ?? {}),
+        },
+      );
+
+    case MapType.petal:
+      return Utils.buildUrl(
+        url: 'petalmaps://poidetail',
+        queryParams: {
+          'marker': '${coords.latitude},${coords.longitude}',
+          'z': '$zoomLevel',
+          ...(extraParams ?? {}),
+        },
+      );
+
+    case MapType.tomtomgo:
+      if (Platform.isIOS) {
+        // currently uses the navigate endpoint on iOS, even when just showing a marker
+        return Utils.buildUrl(
+          url: 'tomtomgo://x-callback-url/navigate',
+          queryParams: {
+            'destination': '${coords.latitude},${coords.longitude}',
+            ...(extraParams ?? {}),
+          },
+        );
+      }
+      return Utils.buildUrl(
+        url: 'geo:${coords.latitude},${coords.longitude}',
+        queryParams: {
+          'q':
+              '${coords.latitude},${coords.longitude}${title != null && title.isNotEmpty ? '($title)' : ''}',
+          ...(extraParams ?? {}),
+        },
+      );
+
+    case MapType.copilot:
+      // Documentation:
+      // https://developer.trimblemaps.com/copilot-navigation/v10-19/feature-guide/advanced-features/url-launch/
+      return Utils.buildUrl(
+        url: 'copilot://mydestination',
+        queryParams: {
+          'type': 'LOCATION',
+          'action': 'VIEW',
+          'marker': '${coords.latitude},${coords.longitude}',
+          'name': title ?? '',
+          ...(extraParams ?? {}),
+        },
+      );
+
+    case MapType.tomtomgofleet:
+      return Utils.buildUrl(
+        url: 'geo:${coords.latitude},${coords.longitude}',
+        queryParams: {
+          'q':
+              '${coords.latitude},${coords.longitude}${title != null && title.isNotEmpty ? '($title)' : ''}',
+          ...(extraParams ?? {}),
+        },
+      );
+
+    case MapType.sygicTruck:
+      // Documentation:
+      // https://www.sygic.com/developers/professional-navigation-sdk/introduction
+      return Utils.buildUrl(
+        url:
+            'com.sygic.aura://coordinate|${coords.longitude}|${coords.latitude}|show',
+        queryParams: {
+          ...(extraParams ?? {}),
+        },
+      );
+
+    case MapType.flitsmeister:
+      if (Platform.isIOS) {
+        return Utils.buildUrl(
+          url: 'flitsmeister://',
+          queryParams: {
+            'geo': '${coords.latitude},${coords.longitude}',
+            ...(extraParams ?? {}),
+          },
+        );
+      }
+      return Utils.buildUrl(
+        url: 'geo:${coords.latitude},${coords.longitude}',
+        queryParams: {
+          'q': '${coords.latitude},${coords.longitude}',
+          ...(extraParams ?? {}),
+        },
+      );
+
+    case MapType.truckmeister:
+      if (Platform.isIOS) {
+        return Utils.buildUrl(
+          url: 'truckmeister://',
+          queryParams: {
+            'geo': '${coords.latitude},${coords.longitude}',
+            ...(extraParams ?? {}),
+          },
+        );
+      }
+      return Utils.buildUrl(
+        url: 'geo:${coords.latitude},${coords.longitude}',
+        queryParams: {
+          'q': '${coords.latitude},${coords.longitude}',
+          ...(extraParams ?? {}),
+        },
+      );
+
+    case MapType.naver:
+      return Utils.buildUrl(
+        url: 'nmap://place',
+        queryParams: {
+          'lat': '${coords.latitude}',
+          'lng': '${coords.longitude}',
+          'zoom': '$zoomLevel',
+          'name': title,
+          ...(extraParams ?? {}),
+        },
+      );
+
+    case MapType.kakao:
+      return Utils.buildUrl(
+        url: 'kakaomap://look',
+        queryParams: {
+          'p': '${coords.latitude},${coords.longitude}',
+          ...(extraParams ?? {}),
+        },
+      );
+
+    case MapType.tmap:
+      return Utils.buildUrl(
+        url: 'tmap://viewmap',
+        queryParams: {
+          'name': '$title',
+          'x': '${coords.longitude}',
+          'y': '${coords.latitude}',
           ...(extraParams ?? {}),
         },
       );