+++ lib/src/models.dart
@@ -1,25 +1,94 @@
 import 'package:map_launcher/src/map_launcher.dart';
 import 'package:map_launcher/src/utils.dart';
 
+/// Defines the map types supported by this plugin
 enum MapType {
+  /// Apple Maps
+  /// Only available on iOS
   apple,
+
+  /// Google Maps
   google,
+
+  /// Google Maps Go
+  /// Only available on Android
   googleGo,
+
+  /// Amap (Gaode Maps)
   amap,
+
+  /// Baidu Maps
   baidu,
+
+  /// Waze
   waze,
+
+  /// Yandex Maps
   yandexMaps,
+
+  /// Yandex Navi
   yandexNavi,
+
+  /// Citymapper
   citymapper,
+
+  /// Maps.me
   mapswithme,
+
+  /// OsmAnd
   osmand,
+
+  /// OsmAnd+
+  /// Only available on Android
   osmandplus,
+
+  /// DoubleGis
   doubleGis,
+
+  /// Tencent (QQ Maps)
   tencent,
+
+  /// HERE WeGo
   here,
+
+  /// Petal Maps
+  /// Only available on Android
+  petal,
+
+  /// TomTom Go
+  tomtomgo,
+
+  /// TomTom Go Fleet
+  tomtomgofleet,
+
+  /// CoPilot
+  copilot,
+
+  /// Sygic Truck
+  sygicTruck,
+
+  /// Flitsmeister
+  /// Only available on Android
+  flitsmeister,
+
+  /// Truckmeister
+  /// Only available on Android
+  truckmeister,
+
+  // Naver Map
+  naver,
+
+  // KakaoMap
+  kakao,
+
+  // TMAP
+  tmap,
+
+  /// MapyCZ
   mapyCz,
 }
 
+/// Defines the supported modes of transportation for [showDirections]
 enum DirectionsMode {
   driving,
   walking,
@@ -27,6 +96,7 @@ enum DirectionsMode {
   bicycling,
 }
 
+/// Class that holds latitude and longitude coordinates
 class Coords {
   final double latitude;
   final double longitude;
@@ -34,6 +104,19 @@ class Coords {
   Coords(this.latitude, this.longitude);
 }
 
+/// Class that holds lat/lng coordinates and optional title
+class Waypoint {
+  final Coords coords;
+  final String? title;
+
+  Waypoint(double latitude, double longitude, [this.title])
+      : coords = Coords(latitude, longitude);
+
+  double get latitude => coords.latitude;
+  double get longitude => coords.longitude;
+}
+
+/// Class that holds all the information needed to launch a map
 class AvailableMap {
   String mapName;
   MapType mapType;
@@ -45,6 +128,7 @@ class AvailableMap {
     required this.icon,
   });
 
+  /// Parses json object to [AvailableMap]
   static AvailableMap? fromJson(json) {
     final MapType? mapType =
         Utils.enumFromString(MapType.values, json['mapType']);
@@ -59,6 +143,7 @@ class AvailableMap {
     }
   }
 
+  /// Launches current map and shows marker at `coords`
   Future<void> showMarker({
     required Coords coords,
     required String title,
@@ -76,12 +161,13 @@ class AvailableMap {
     );
   }
 
+  /// Launches current map and shows directions to `destination`
   Future<void> showDirections({
     required Coords destination,
     String? destinationTitle,
     Coords? origin,
     String? originTitle,
-    List<Coords>? waypoints,
+    List<Waypoint>? waypoints,
     DirectionsMode directionsMode = DirectionsMode.driving,
     Map<String, String>? extraParams,
   }) {